showcase bug with expo-blur and expo-router on android
Bug 1:
if you render a blur view and use a modal screen the modal screen at one
points stops showing up. The screen is rendered, but not visible.
If the screen would have for example a button - you would not see the
button, but you will be able to see it. If you were to change the uimode
with
 adb shell "cmd uimode night yes|no”

then the modal will render properly.

Bug 2:
The modal view becomes semi-transparent. You cannot place a div that is
non-transparent anymore. If there is no BlurView, then it works fine.

update expo to latest versions